+++ tests/Controller/WebhookControllerTest.php
@@ -13,6 +13,7 @@ namespace BoShurik\TelegramBotBundle\Tests\Controller;
 
 use BoShurik\TelegramBotBundle\Controller\WebhookController;
 use BoShurik\TelegramBotBundle\Event\WebhookEvent;
+use BoShurik\TelegramBotBundle\Messenger\TelegramMessage;
 use BoShurik\TelegramBotBundle\Telegram\Telegram;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
@@ -20,6 +21,8 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\MessageBusInterface;
 use TelegramBot\Api\Types\Update;
 
 class WebhookControllerTest extends TestCase
@@ -35,9 +38,19 @@ class WebhookControllerTest extends TestCase
     private $eventDispatcher;
 
     /**
+     * @var MessageBusInterface|MockObject
+     */
+    private $bus;
+
+    /**
      * @var WebhookController
      */
     private $controller;
+
+    /**
+     * @var WebhookController
+     */
+    private $controllerWithBus;
 
     /**
      * @inheritDoc
@@ -46,7 +59,9 @@ class WebhookControllerTest extends TestCase
     {
         $this->telegram = $this->createMock(Telegram::class);
         $this->eventDispatcher = $this->createMock(EventDispatcherInterface::class);
+        $this->bus = $this->createMock(MessageBusInterface::class);
         $this->controller = new WebhookController($this->telegram, $this->eventDispatcher);
+        $this->controllerWithBus = new WebhookController($this->telegram, $this->eventDispatcher, $this->bus);
     }
 
     public function testEmptyData(): void
@@ -80,8 +95,11 @@ class WebhookControllerTest extends TestCase
                 if ($event->getRequest() !== $request) {
                     return false;
                 }
+                if (!$event->getUpdate() instanceof Update) {
+                    return false;
+                }
 
-                return $event->getUpdate() instanceof Update;
+                return $event->getUpdate()->getUpdateId() === 0;
             }))
             ->willReturnCallback(function (WebhookEvent $event) {
                 return $event;
@@ -112,6 +130,9 @@ class WebhookControllerTest extends TestCase
                 if ($event->getRequest() !== $request) {
                     return false;
                 }
+                if (!$event->getUpdate() instanceof Update) {
+                    return false;
+                }
 
                 return $event->getUpdate()->getUpdateId() === 0;
             }))
@@ -125,6 +146,51 @@ class WebhookControllerTest extends TestCase
         $response = $this->controller->indexAction($request);
 
         $this->assertSame($expectedResponse, $response);
+    }
+
+    public function testBus()
+    {
+        $request = $this->createRequest(json_encode([
+            'update_id' => 0,
+        ]));
+
+        $this->telegram
+            ->expects($this->never())
+            ->method('processUpdate')
+        ;
+        $this->bus
+            ->expects($this->once())
+            ->method('dispatch')
+            ->with($this->callback(function ($message) {
+                if (!$message instanceof TelegramMessage) {
+                    return false;
+                }
+
+                return $message->getUpdate()->getUpdateId() === 0;
+            }))
+            ->willReturn(new Envelope(new \stdClass()));
+
+        $this->eventDispatcher
+            ->expects($this->once())
+            ->method('dispatch')
+            ->with($this->callback(function ($event) use ($request) {
+                if (!$event instanceof WebhookEvent) {
+                    return false;
+                }
+                if ($event->getRequest() !== $request) {
+                    return false;
+                }
+                if (!$event->getUpdate() instanceof Update) {
+                    return false;
+                }
+
+                return $event->getUpdate()->getUpdateId() === 0;
+            }))
+            ->willReturnCallback(function (WebhookEvent $event) {
+                return $event;
+            });
+
+        $this->controllerWithBus->indexAction($request);
     }
 
     private function createRequest(string $content): Request